:sparkles: Black latest rev from master :sparkles:

+++ databricks/koalas/tests/test_stats.py
@@ -23,47 +23,55 @@ from databricks.koalas.testing.utils import ReusedSQLTestCase, SQLTestUtils
 
 
 class StatsTest(ReusedSQLTestCase, SQLTestUtils):
-
     def _test_stat_functions(self, pdf, kdf):
-        functions = ['max', 'min', 'mean', 'sum']
+        functions = ["max", "min", "mean", "sum"]
         for funcname in functions:
             self.assert_eq(getattr(kdf.A, funcname)(), getattr(pdf.A, funcname)())
             self.assert_eq(getattr(kdf, funcname)(), getattr(pdf, funcname)())
 
-        functions = ['std', 'var']
+        functions = ["std", "var"]
         for funcname in functions:
-            self.assert_eq(getattr(kdf.A, funcname)(), getattr(pdf.A, funcname)(), almost=True)
-            self.assert_eq(getattr(kdf, funcname)(), getattr(pdf, funcname)(), almost=True)
+            self.assert_eq(
+                getattr(kdf.A, funcname)(), getattr(pdf.A, funcname)(), almost=True
+            )
+            self.assert_eq(
+                getattr(kdf, funcname)(), getattr(pdf, funcname)(), almost=True
+            )
 
         # NOTE: To test skew, kurt, and median, just make sure they run.
         #       The numbers are different in spark and pandas.
-        functions = ['skew', 'kurt', 'median']
+        functions = ["skew", "kurt", "median"]
         for funcname in functions:
             getattr(kdf.A, funcname)()
             getattr(kdf, funcname)()
 
     def test_stat_functions(self):
-        pdf = pd.DataFrame({'A': [1, 2, 3, 4],
-                            'B': [1.0, 2.1, 3, 4]})
+        pdf = pd.DataFrame({"A": [1, 2, 3, 4], "B": [1.0, 2.1, 3, 4]})
         kdf = ks.from_pandas(pdf)
         self._test_stat_functions(pdf, kdf)
 
     def test_stat_functions_multiindex_column(self):
-        arrays = [np.array(['A', 'A', 'B', 'B']),
-                  np.array(['one', 'two', 'one', 'two'])]
-        pdf = pd.DataFrame(np.random.randn(3, 4), index=['A', 'B', 'C'], columns=arrays)
+        arrays = [
+            np.array(["A", "A", "B", "B"]),
+            np.array(["one", "two", "one", "two"]),
+        ]
+        pdf = pd.DataFrame(np.random.randn(3, 4), index=["A", "B", "C"], columns=arrays)
         kdf = ks.from_pandas(pdf)
         self._test_stat_functions(pdf, kdf)
 
     def test_abs(self):
-        pdf = pd.DataFrame({'A': [1, -2, 3, -4, 5],
-                            'B': [1., -2, 3, -4, 5],
-                            'C': [-6., -7, -8, -9, 10],
-                            'D': ['a', 'b', 'c', 'd', 'e']})
+        pdf = pd.DataFrame(
+            {
+                "A": [1, -2, 3, -4, 5],
+                "B": [1.0, -2, 3, -4, 5],
+                "C": [-6.0, -7, -8, -9, 10],
+                "D": ["a", "b", "c", "d", "e"],
+            }
+        )
         kdf = ks.from_pandas(pdf)
         self.assert_eq(kdf.A.abs(), pdf.A.abs())
         self.assert_eq(kdf.B.abs(), pdf.B.abs())
-        self.assert_eq(kdf[['B', 'C']].abs(), pdf[['B', 'C']].abs())
+        self.assert_eq(kdf[["B", "C"]].abs(), pdf[["B", "C"]].abs())
         # self.assert_eq(kdf.select('A', 'B').abs(), pdf[['A', 'B']].abs())
 
     def test_axis_on_dataframe(self):
@@ -72,11 +80,15 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
         # Less than 'compute.shortcut_limit' will execute a shortcut
         # by using collected pandas dataframe directly.
         # now we set the 'compute.shortcut_limit' as 1000 explicitly
-        with option_context('compute.shortcut_limit', 1000):
-            pdf = pd.DataFrame({'A': [1, -2, 3, -4, 5] * 300,
-                                'B': [1., -2, 3, -4, 5] * 300,
-                                'C': [-6., -7, -8, -9, 10] * 300,
-                                'D': [True, False, True, False, False] * 300})
+        with option_context("compute.shortcut_limit", 1000):
+            pdf = pd.DataFrame(
+                {
+                    "A": [1, -2, 3, -4, 5] * 300,
+                    "B": [1.0, -2, 3, -4, 5] * 300,
+                    "C": [-6.0, -7, -8, -9, 10] * 300,
+                    "D": [True, False, True, False, False] * 300,
+                }
+            )
             kdf = ks.from_pandas(pdf)
             self.assert_eq(kdf.count(axis=1), pdf.count(axis=1))
             self.assert_eq(kdf.var(axis=1), pdf.var(axis=1))
@@ -90,7 +102,7 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
 
     def test_corr(self):
         # Disable arrow execution since corr() is using UDT internally which is not supported.
-        with self.sql_conf({'spark.sql.execution.arrow.enabled': False}):
+        with self.sql_conf({"spark.sql.execution.arrow.enabled": False}):
             # DataFrame
             # we do not handle NaNs for now
             pdf = pd.util.testing.makeMissingDataframe(0.3, 42).fillna(0)
@@ -108,38 +120,43 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
             self.assertRaises(TypeError, lambda: kser_a.corr(kdf))
 
             # multi-index columns
-            columns = pd.MultiIndex.from_tuples([('X', 'A'), ('X', 'B'), ('Y', 'C'), ('Z', 'D')])
+            columns = pd.MultiIndex.from_tuples(
+                [("X", "A"), ("X", "B"), ("Y", "C"), ("Z", "D")]
+            )
             pdf.columns = columns
             kdf.columns = columns
 
             self.assert_eq(kdf.corr(), pdf.corr(), almost=True)
 
             # Series
-            pser_xa = pdf[('X', 'A')]
-            pser_xb = pdf[('X', 'B')]
-            kser_xa = kdf[('X', 'A')]
-            kser_xb = kdf[('X', 'B')]
+            pser_xa = pdf[("X", "A")]
+            pser_xb = pdf[("X", "B")]
+            kser_xa = kdf[("X", "A")]
+            kser_xb = kdf[("X", "B")]
 
             self.assertAlmostEqual(kser_xa.corr(kser_xb), pser_xa.corr(pser_xb))
 
     def test_cov_corr_meta(self):
         # Disable arrow execution since corr() is using UDT internally which is not supported.
-        with self.sql_conf({'spark.sql.execution.arrow.enabled': False}):
-            pdf = pd.DataFrame({'a': np.array([1, 2, 3], dtype='i1'),
-                                'b': np.array([1, 2, 3], dtype='i2'),
-                                'c': np.array([1, 2, 3], dtype='i4'),
-                                'd': np.array([1, 2, 3]),
-                                'e': np.array([1.0, 2.0, 3.0], dtype='f4'),
-                                'f': np.array([1.0, 2.0, 3.0]),
-                                'g': np.array([True, False, True]),
-                                'h': np.array(list('abc'))},
-                               index=pd.Index([1, 2, 3], name='myindex'))
+        with self.sql_conf({"spark.sql.execution.arrow.enabled": False}):
+            pdf = pd.DataFrame(
+                {
+                    "a": np.array([1, 2, 3], dtype="i1"),
+                    "b": np.array([1, 2, 3], dtype="i2"),
+                    "c": np.array([1, 2, 3], dtype="i4"),
+                    "d": np.array([1, 2, 3]),
+                    "e": np.array([1.0, 2.0, 3.0], dtype="f4"),
+                    "f": np.array([1.0, 2.0, 3.0]),
+                    "g": np.array([True, False, True]),
+                    "h": np.array(list("abc")),
+                },
+                index=pd.Index([1, 2, 3], name="myindex"),
+            )
             kdf = ks.from_pandas(pdf)
             self.assert_eq(kdf.corr(), pdf.corr())
 
     def test_stats_on_boolean_dataframe(self):
-        pdf = pd.DataFrame({'A': [True, False, True],
-                            'B': [False, False, True]})
+        pdf = pd.DataFrame({"A": [True, False, True], "B": [False, False, True]})
         kdf = ks.from_pandas(pdf)
 
         pd.testing.assert_series_equal(kdf.min(), pdf.min())
@@ -165,9 +182,9 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
         self.assertAlmostEqual(kser.std(), pser.std())
 
     def test_some_stats_functions_should_discard_non_numeric_columns_by_default(self):
-        pdf = pd.DataFrame({'i': [0, 1, 2],
-                            'b': [False, False, True],
-                            's': ['x', 'y', 'z']})
+        pdf = pd.DataFrame(
+            {"i": [0, 1, 2], "b": [False, False, True], "s": ["x", "y", "z"]}
+        )
         kdf = ks.from_pandas(pdf)
 
         # min and max do not discard non-numeric columns by default
@@ -185,29 +202,47 @@ class StatsTest(ReusedSQLTestCase, SQLTestUtils):
         self.assertEqual(len(kdf.kurtosis()), len(pdf.kurtosis()))
         self.assertEqual(len(kdf.skew()), len(pdf.skew()))
 
-    def test_stats_on_non_numeric_columns_should_be_discarded_if_numeric_only_is_true(self):
-        pdf = pd.DataFrame({'i': [0, 1, 2],
-                            'b': [False, False, True],
-                            's': ['x', 'y', 'z']})
+    def test_stats_on_non_numeric_columns_should_be_discarded_if_numeric_only_is_true(
+        self,
+    ):
+        pdf = pd.DataFrame(
+            {"i": [0, 1, 2], "b": [False, False, True], "s": ["x", "y", "z"]}
+        )
         kdf = ks.from_pandas(pdf)
 
-        self.assertEqual(len(kdf.sum(numeric_only=True)), len(pdf.sum(numeric_only=True)))
-        self.assertEqual(len(kdf.mean(numeric_only=True)), len(pdf.mean(numeric_only=True)))
+        self.assertEqual(
+            len(kdf.sum(numeric_only=True)), len(pdf.sum(numeric_only=True))
+        )
+        self.assertEqual(
+            len(kdf.mean(numeric_only=True)), len(pdf.mean(numeric_only=True))
+        )
 
-        self.assertEqual(len(kdf.var(numeric_only=True)), len(pdf.var(numeric_only=True)))
-        self.assertEqual(len(kdf.std(numeric_only=True)), len(pdf.std(numeric_only=True)))
+        self.assertEqual(
+            len(kdf.var(numeric_only=True)), len(pdf.var(numeric_only=True))
+        )
+        self.assertEqual(
+            len(kdf.std(numeric_only=True)), len(pdf.std(numeric_only=True))
+        )
 
-        self.assertEqual(len(kdf.kurtosis(numeric_only=True)), len(pdf.kurtosis(numeric_only=True)))
-        self.assertEqual(len(kdf.skew(numeric_only=True)), len(pdf.skew(numeric_only=True)))
+        self.assertEqual(
+            len(kdf.kurtosis(numeric_only=True)), len(pdf.kurtosis(numeric_only=True))
+        )
+        self.assertEqual(
+            len(kdf.skew(numeric_only=True)), len(pdf.skew(numeric_only=True))
+        )
 
-    def test_stats_on_non_numeric_columns_should_not_be_discarded_if_numeric_only_is_false(self):
-        pdf = pd.DataFrame({'i': [0, 1, 2],
-                            'b': [False, False, True],
-                            's': ['x', 'y', 'z']})
+    def test_stats_on_non_numeric_columns_should_not_be_discarded_if_numeric_only_is_false(
+        self,
+    ):
+        pdf = pd.DataFrame(
+            {"i": [0, 1, 2], "b": [False, False, True], "s": ["x", "y", "z"]}
+        )
         kdf = ks.from_pandas(pdf)
 
         # the lengths are the same, but the results are different.
-        self.assertEqual(len(kdf.sum(numeric_only=False)), len(pdf.sum(numeric_only=False)))
+        self.assertEqual(
+            len(kdf.sum(numeric_only=False)), len(pdf.sum(numeric_only=False))
+        )
 
         # pandas fails belows.
         # self.assertEqual(len(kdf.mean(numeric_only=False)), len(pdf.mean(numeric_only=False)))
